Readded original function for db upgrade and moved to alternative approach to db upgrade_alt
