+++ chia/cmds/db.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 import click
 from chia.cmds.db_upgrade_func import db_upgrade_func
+from chia.cmds.db_upgrade_alt_func import db_upgrade_alt_func
 
 
 @click.group("db", short_help="Manage the blockchain database")
@@ -9,6 +10,29 @@ def db_cmd() -> None:
 
 
 @db_cmd.command("upgrade", short_help="EXPERIMENTAL: upgrade a v1 database to v2 out of place")
+@click.option("--input", default=None, type=click.Path(), help="specify input database file")
+@click.option("--output", default=None, type=click.Path(), help="specify output database file")
+@click.option(
+    "--no-update-config",
+    default=False,
+    is_flag=True,
+    help="don't update config file to point to new database. When specifying a "
+    "custom output file, the config will not be updated regardless",
+)
+@click.pass_context
+def db_upgrade_cmd(ctx: click.Context, no_update_config: bool, **kwargs) -> None:
+
+    in_db_path = kwargs.get("input")
+    out_db_path = kwargs.get("output")
+    db_upgrade_func(
+        Path(ctx.obj["root_path"]),
+        None if in_db_path is None else Path(in_db_path),
+        None if out_db_path is None else Path(out_db_path),
+        no_update_config,
+    )
+
+
+@db_cmd.command("upgrade_alt", short_help="EXPERIMENTAL: upgrade v1 database to v2 - alternative approach")
 @click.option("--input", default=None, type=click.Path(), help="specify input database file")
 @click.option("--output", default=None, type=click.Path(), help="specify output database file")
 @click.option("--hdd", default=False, is_flag=True, help="select if database is located on HDD device. (DEFAULT=false)")
@@ -38,17 +62,14 @@ def db_cmd() -> None:
     "custom output file, the config will not be updated regardless",
 )
 @click.pass_context
-def db_upgrade_cmd(
+def db_upgrade_alt_cmd(
     ctx: click.Context, no_update_config: bool, offline: bool, hdd: bool, check_only: bool, **kwargs
 ) -> None:
 
     in_db_path = kwargs.get("input")
     out_db_path = kwargs.get("output")
-    #    hdd = kwargs.get("hdd")
-    #    offline = kwargs.get("offline")
     temp_path = kwargs.get("temp_store_path")
-    #    check_only = kwargs.get("check_only")
-    db_upgrade_func(
+    db_upgrade_alt_func(
         Path(ctx.obj["root_path"]),
         None if in_db_path is None else Path(in_db_path),
         None if out_db_path is None else Path(out_db_path),
@@ -58,9 +79,3 @@ def db_upgrade_cmd(
         None if temp_path is None else Path(temp_path),
         check_only,
     )
-
-
-if __name__ == "__main__":
-    from chia.util.default_root import DEFAULT_ROOT_PATH
-
-    db_upgrade_func(DEFAULT_ROOT_PATH)
